Refactor file into two seprarate files
[MAILPOET-3222]
